+++ src/com/geoscope/GeoEye/Space/Server/TGeoScopeServerInfo.java
@@ -91,7 +91,7 @@ public class TGeoScopeServerInfo {
 			throw new Exception("User is not initialized"); //. =>
 		String URL1 = Server.Address;
 		//. add command path
-		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Integer.toString(Server.User.UserID);
+		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Long.toString(Server.User.UserID);
 		String URL2 = "SpaceServers.xml";
 		//. add command parameters
 		URL2 = URL2+"?"+"1"/*command version*/;
